requirements.txt
google-auth>=2.6
requests>=2.27
beautifulsoup4
google-api-python-client
google-auth-oauthlib
google-cloud-aiplatform
Flask==3.0.0
google-cloud-firestore==2.13.1
gunicorn==21.2.0